+++ packages/recorder/src/index.ts
@@ -7,7 +7,7 @@ const isElectronProcess = require('is-electron');
 let vscodeEnv;
 try {
   // eslint-disable-next-line
-  const vscode = require('is-electron');
+  const vscode = require('vscode');
   vscodeEnv = vscode.env;
 } catch (error) {
   // ignore
@@ -104,7 +104,6 @@ export async function record(originParam: IGoldlogParam) {
 }
 
 export function recordDAU() {
-  console.log('recorder[dau]');
   return recordUV({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,
@@ -118,74 +117,42 @@ export function recordDAU() {
 
 // call in DefinitionProvider
 export function recordDefinitionProvider() {
-  console.log('recorder[definitionProvider]');
   return recordUV({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,
     action: 'definitionProvider',
-    data: {
-      platform: process.platform,
-      locale: vscodeEnv ? vscodeEnv.language : 'zh-CN',
-    },
   });
 }
 
 // call in HoverProvider
 export function recordHoverProvider() {
-  console.log('recorder[hoverProvider]');
   return recordUV({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,
     action: 'hoverProvider',
-    data: {
-      platform: process.platform,
-      locale: vscodeEnv ? vscodeEnv.language : 'zh-CN',
-    },
   });
 }
 
 // call in CompletionItemProvider
 export function recordCompletionItemProvider() {
-  console.log('recorder[completionItemProvider]');
   return recordUV({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,
     action: 'completionItemProvider',
-    data: {
-      platform: process.platform,
-      locale: vscodeEnv ? vscodeEnv.language : 'zh-CN',
-    },
   });
 }
 
 // call in completion item select commands
 export function recordCompletionItemSelect() {
-  console.log('recorder[completionItemSelect]');
   return recordUV({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,
     action: 'completionItemSelect',
-    data: {
-      platform: process.platform,
-      locale: vscodeEnv ? vscodeEnv.language : 'zh-CN',
-    },
-  });
-}
-
-export function recordExecuteCommand(command: string) {
-  console.log('recorder[recordExecuteCommand]');
-  return record({
-    namespace: MAIN_KEY,
-    module: RECORD_MODULE_KEY,
-    action: 'executeCommand',
-    data: {
-      command,
-    },
   });
 }
 
 // active a extension
-export function recordActivate(data: { extension: string; version?: string }) {
+function recordActivate(data: { extension: string; version?: string }) {
   return record({
     namespace: MAIN_KEY,
     module: RECORD_MODULE_KEY,