+++ extensions/iceworks-config-helper/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.3.6
+
+- fix: compatible with scenarios without material source
+
+## 0.3.5
+
+- fix: if there is an active window, then do not show the welcome page
+
 ## 0.3.4
 
 - fix: show welcome page many times