+++ extensions/iceworks-doctor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.1.3
+
+- fix: compatible with scenarios without material source
+
+## 0.1.2
+
+- fix: if there is an active window, then do not show the welcome page
+
 ## 0.1.1
 
 - docs: add logo