+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.3.14
+
+- feat: custom material sources will be at first in the material sources list
+
+## 0.3.13
+
+- fix: if there is an active window, then do not show the welcome page
+
 ## 0.3.12
 
 - fix: show welcome page many times