+++ extensions/iceworks-style-helper/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.4.0
+
+- feat: support completion in vue class='xxx'
+
+## 0.3.11
+
+- fix: compatible with scenarios without material source
+
+## 0.3.10
+
+- fix: if there is an active window, then do not show the welcome page
+
 ## 0.3.9
 
 - feat: Add welcome page