+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.1.18
+
+- feat: add code editor for page template
+- fix: no select any blocks prompt message is not correct
+
+## 0.1.17
+
+- fix: if there is an active window, then do not show the welcome page
+
 ## 0.1.16
 
 - fix: show welcome page many times